Change the verification request
Please change the verification request to the one that wouldn't cause changes to where we send the data (which would lead to the token owner losing that data). The getStreamWidgetUrl would work for all tokens, including the ones generated through our APIs that can be prohibited from accessing the setCallbackUrl method. getStreamWidgetUrl is always guaranteed to work if a functioning token is used.